Upgrading Plugin to Latest Gradle and Jasper
Upgrading the plugin to use a more recent version of gradle, v5.6.2, and
jasperreports, v6.9.0. As a result, moved away from the bintray
publishing plugin to instead use the gradle publishing plugin.

Moving to the gradle publishing plugin has several good & bad side
effects. The good was that it eliminated all the bintray tasks and
configuration and automatically generates the
src/main/resources/META-INF/gradle-plugins *.property file based on the
plugin id in the publishing plugin configuration. The bad is that the
author of the plugin will need to do a little work to use the "new" way
of publishing plugins to the gradle plugin portal as described in the
following link: https://plugins.gradle.org/docs/submit. Note that link
also has instructions at the top of the page on how to migrate
publishing from bintray.

Removing mavenLocal() from main repositories section